Initial commit.
Get basic socket.io server working and tested. Currently have to have
a bootstrap.js that requires server.js and tells express to listen
because of the need to do the same in specs. Not sure if this is the
best way to go about this or not, certainly worth looking into
eventually.